+++ mods/BrewLAN/lua/FieldEngineers.lua
@@ -6,9 +6,10 @@ function SCUFieldUpgrade(SuperClass)
         OnCreate = function(self)
             SuperClass.OnCreate(self)
             if self:GetAIBrain().BrainType == 'Human' then
-                self:AddBuildRestriction(categories.BUILTBYTIER3FIELD)
                 if categories.BUILTBYFIELD then
                     self:AddBuildRestriction(categories.BUILTBYFIELD)
+                else
+                    self:AddBuildRestriction(categories.BUILTBYTIER3FIELD)
                 end
             end
         end,
@@ -16,23 +17,36 @@ function SCUFieldUpgrade(SuperClass)
         CreateEnhancement = function(self, enh)
             SuperClass.CreateEnhancement(self, enh)
             local bp = self:GetBlueprint().Enhancements[enh]
+            local ParseEntityCategory = ParseEntityCategory
             if not bp then return end
-            --Field
-            local restriction = (categories.ECONOMIC * categories.TECH3) - categories.BUILTBYTIER3FIELD - categories.MASSSTORAGE - categories.ENERGYSTORAGE - categories.HYDROCARBON
+            --Prevent ED4 ED5 duplication
+            local restriction = categories.urb4206
+            --Prevent duplication of advanced resouorce buildings.
+            for i, v in {'a', 'e', 'r', 's'} do
+                if categories['s'..v..'b1311'] then
+                    restriction = restriction + ParseEntityCategory('u'..v..'b1301')
+                end
+                if categories['s'..v..'b1312'] then
+                    restriction = restriction + ParseEntityCategory('u'..v..'b1302')
+                end
+                if categories['s'..v..'b1313'] then
+                    restriction = restriction + ParseEntityCategory('u'..v..'b1303')
+                end
+            end
             if enh == 'Field' then
-                self:RemoveBuildRestriction(categories.BUILTBYTIER3FIELD)
                 if categories.BUILTBYFIELD then
                     self:RemoveBuildRestriction(categories.BUILTBYFIELD)
+                else
+                    self:RemoveBuildRestriction(categories.BUILTBYTIER3FIELD)
                 end
                 self:AddBuildRestriction(restriction)
-                self:AddBuildRestriction(categories.urb4206)
             elseif enh == 'FieldRemove' then
-                self:AddBuildRestriction(categories.BUILTBYTIER3FIELD)
                 if categories.BUILTBYFIELD then
                     self:AddBuildRestriction(categories.BUILTBYFIELD)
+                else
+                    self:AddBuildRestriction(categories.BUILTBYTIER3FIELD)
                 end
                 self:RemoveBuildRestriction(restriction)
-                self:RemoveBuildRestriction(categories.urb4206)
             end
         end,
     }