+++ mods/BrewLAN/units/SRL0209/SRL0209_script.lua
@@ -1,27 +1,20 @@
-#****************************************************************************
-#**
-#**  Summary  :  Cybran T2 Combat Engineer Script
-#**
-#****************************************************************************
-
+--------------------------------------------------------------------------------
+--  Summary  :  Cybran Combat Engineer Script
+--------------------------------------------------------------------------------
 local CConstructionUnit = import('/lua/cybranunits.lua').CConstructionUnit
 local CDFElectronBolterWeapon = import('/lua/cybranweapons.lua').CDFElectronBolterWeapon
-
+--------------------------------------------------------------------------------
 SRL0209 = Class(CConstructionUnit) {
-
     Weapons = {
         Bolter = Class(CDFElectronBolterWeapon) {},
     },
 
-    Treads = {
-        ScrollTreads = true,
-        BoneName = 'BRL0209',
-        TreadMarks = 'tank_treads_albedo',
-        TreadMarksSizeX = 0.65,
-        TreadMarksSizeZ = 0.4,
-        TreadMarksInterval = 0.3,
-        TreadOffset = { 0, 0, 0 },
-    },
+    OnCreate = function(self)
+        CConstructionUnit.OnCreate(self)
+        if self:GetAIBrain().BrainType == 'Human' then
+            self:AddBuildRestriction(categories.urb4206)
+        end
+    end,
 
     OnStopBeingBuilt = function(self, ...)
         self:SetMaintenanceConsumptionActive()
@@ -33,21 +26,22 @@ SRL0209 = Class(CConstructionUnit) {
         self:SetWeaponEnabledByLabel('Bolter', false)
         CConstructionUnit.OnStartBuild(self, unitBeingBuilt, order)
     end,
-    
+
     OnStopBuild = function(self)
         --Re-enable the gun after done building
         self:SetWeaponEnabledByLabel('Bolter', true)
         CConstructionUnit.OnStopBuild(self)
     end,
-    
+
     OnStartReclaim = function(self, target)
         CConstructionUnit.OnStartReclaim(self, target)
         self:SetAllWeaponsEnabled(false)
     end,
-    
+
     OnStopReclaim = function(self, target)
         CConstructionUnit.OnStopReclaim(self, target)
         self:SetAllWeaponsEnabled( true)
     end,
 }
+
 TypeClass = SRL0209